multiple updates
Changed the password in create account to only show dots to hide the password from prying eyes.
Added a shadow around the add user informaiton fill in blocks.
Moved the nav bar from the base.html so it can show different items depending on the page it is showing.
Created a change password page.
Activated links on the login page for create account and change password.