fixed registration page and set secret key
On the config.py file, I added a note of what I set the secret key to and how I set the secret key so it would not show up as plain text in the code. I have also got the registration page working again and the password is hashed, so nobody can read it if they get into the database.