Updated the look of welcome page, and register.
The register page now accepts entries and will let you know if the input was a success or not. Added a user home page for when I get the login working. I hope to have that done tomorrow. I need to remove a lot of the data in the database. Passwords are also hashed.