created a 2nd table and set up foreign keys.
The second table was created to have one created. We can change the columns later depending on how you want the questions and answers to be saved in the database. I've been watching Corey Schafer's videos for the Python Flask Tutorial: Full-Featured Web App series of about 15 videos. I'm up to video 5.