Answer checker can now verify if your question is right or wrong.
The answer checker can now determine if the problem the user put in is correct or not and will post the correct answer on the next page. Still need to make the page look better and show a smiley face or something to let them know if they were right or wrong. Still need to add in the multiple chances.