Working on changing password
Have the change password page to populate and it checks the database to verify if the username currently exists. If it exists, will send a message in the control panel that it exists and to change in the database. Currently working on the code to change the password in the database. If username does not exists, a flash message will appear to state it does not exist.