Updated Memory Bank and Memory Bank Game
Changed the current Memory bank to the memory bank input and created a page for the memory bank game. The memory bank input will allow the user or parent to input equations into the database. The memory bank game will work on taking those questions and populating them randomly to the user playing the game.