Changed operator to math_op
Changed everything that said operator for the operator column to math_op. Changed the operator column in the database to math_op since the memory bank input crashed after compiling it as a stand-alone app. Not all of the items from the memory_bank table is populating on the screen.